requirements.txt
annotated-types==0.7.0
anyio==4.6.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cryptography==44.0.2
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
numpy==2.2.3
outcome==1.3.0.post0
pandas==2.2.3
pdfminer.six==20240706
playwright==1.50.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyee==12.1.1
requests>=2.31.0
python-dotenv>=0.19.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.39.2
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.31.1
websocket-client==1.8.0
wsproto==1.2.0
